load-test/requirements.txt
clickhouse-connect
locust~=2.41.6
pyyaml~=6.0.3
requests~=2.32.5
gevent~=25.5.1